Script to generate SQL to add or remove triggers for all tables
This script takes a list of tables as input, and creates two SQL
files: one to create a central table to log all changes to other
tables, and a list of triggers, one for every table in the schema.

The other SQL file will drop everything created by the first script
(i.e. "undo" everything).