Experimenting with a factory function
A factory function that returns a class, not an instance of a class. I
wanted to work out possible ways of adding mixin classes to the class
that's created by the factory function.

There might be something broken about this design; perhaps it's the
blending of closures and class definitions. This doesn't strike me as
the canonical way to modify a class at runtime (a simple class
hierarchy would probably do). Using metaclasses would make the code
unreadable.